requirement.txt
numpy
pandas
python-dotenv
mysql-connector-python
pyodbc
scikit-learn
#-e .